Switch to static ffmpeg 4.2.2 for reliability (#332)
Previously we linked to shared-dll ffmpeg builds.
When downloading the shared builds, some users fail to extract the DLLs,
causing ffmpeg to not work.
Static has a larger filesize, but is less likely to confuse users.

Previously, we linked to the latest ffmpeg dev build.
I'm unsure if ffmpeg dev builds are unstable
or may have bugs varying from version to version.
And there is no URL on zeranoe's website to get the latest release build.
So instead hard-code/pin the current latest release,
and update it from time to time.