Fix longstanding crash when prefs.yaml is corrupted
Instead, warn the user with a popup dialog, and reset settings to
default.

Every few months, a user would report corrscope failing to open
due to a corrupted prefs.yaml.
This commit should stop all occurrences of that from now on.
The next commit will help ensure that we never save corrupted files
in the first place.