Atomically save prefs.yaml to prevent file corruption
This adds a dependency on the atomicwrites package. Note that it was
already a transitive dev-dependency, but is now a direct dependency as
well.

On Linux, atomicwrites fsyncs the directory after saving. This is
unnecessary for consistency after a system crash, and slows down program
shutdown slightly. But I'm too lazy to vendor or rewrite the library to
not fsync the directory.