Replace edge kernel with running sum
Previously we got triggering errors where instead of edge triggering
scores being decided by the waveform around the trigger point, it was
influenced by distant parts of the waveform moving across the wide
sloped portion of the edge detector kernel.

This commit replaces the edge kernel with a running sum, which acts like
an infinitely wide step kernel that the input data slides *within*. As a
result, the difference between the edge scores of two trigger points is
solely determined by the data between those points. (The step kernel and
history buffer are unaffected and still slide within the data.)

This makes triggering less likely to fail. However the global scoring
system sometimes prioritizes bad edges in good regions of the document.
This will be reexamined later.