Use regular trigger's smoothed mean for post triggering
This allows setting mean responsiveness to 0 to affect post triggering
as well as regular (good), without introducing DC offset to the buffer
(which would cause problems).

Additionally it causes the post trigger to use the same DC offset as the
regular trigger, rather than the offset around the trigger point (which
is debatable).