Fix stray triggering results in long areas of audio with no rising edges
When buffer_strength = 0 and no rising edges were in range (based on the
estimated period), corrscope would erroneously return the leftmost
point in its view (based on the estimated period). This would often
result in triggering in the *middle* of a positive segment.

Instead return the center point. In NES oscilloscopes, this returns a
value past the falling edge and in a near-silent region (which is more
visually appealing).