Add "Reload Font List" menu item to fix missing fonts
Matplotlib stores a cache of installed fonts, and does not reload it
when the user installs a new font and tries to use it. We workaround
this by adding a menu item which clears the font cache
(causing Matplotlib to reload its font list). This is easier than
telling users to find and delete a folder on their system.

Also change Trigger tab accelerator so users can use Alt+T to access
Tools -> Reload Font List.