Don't move render filetype dropdown if user types a different extension
The render dialog has a selected filetype ("filter") dropdown to pick
what format to save to by default. This PR changes the code to remember
the selected filetype between renders.

Previously we changed the selected filetype if the user types an
extension belonging to a different type in the list, even if they didn't
pick an extension in the list. This can be confusing, so I removed the
code responsible (instead only remembering the most recently selected
list item).